microservices/helloWorld/requirement.txt
grpcio==1.60.0
grpcio-tools==1.60.0
numpy==1.26.2
protobuf==4.25.1